Python build automation looks in /Users/Shared/EpicGames/
Because there still to be problems with building on the Mac
when there is a space in the UE4 engine path (see issue #61),
the easiest fix is to install without the space. This change
simply adds this path to the end of the mac_candidates list
for automatically finding the UE4 directory.